Fixed brush bugs, refactored to follow best practices, and updated to C++23 - Moved app.SetMenubarCallback() into ApplicationLayer - Removed system classes Init() functions and just used the constructor instead (now possible since classes are more independent) - Simplified entrypoint - Added ToolSettingsController to centralise restrictions like changing the brush width with keys '[' and ']' staying within the bounds 1-256 - Fixed pointer ownership - Moved some things to compile-time evaluation instead of runtime - Updated to C++23 - Made brush compute shader snap its calculations to the pixel grid. This solved gaps in small brush-size tools, pixels only drawing when the top left quadrant of them were clicked (while zoomed in), and made tool radius size more accurate. - Antialiasing needs proper fixing on its own branch so it doesn't accumulate colour over multiple frames and brush smoothing steps

+++ Stylus/src/Tools/ToolData.h
@@ -11,7 +11,7 @@ namespace Stylus {
 	{
 		std::string Name;
 		std::filesystem::path ShaderPath{};
-		size_t PushConstantStructSize{};
+		uint32_t PushConstantStructSize{};
 		Walnut::KeyCode KeyShortcut = Walnut::KeyCode::None;
 	};
